Added actual exception handling
This makes output cleaner in the event of the (possibly many!) IO
exceptions that may occur during use of this program. We have also added
a custom Exception instance to model a couple of logic failures specific
to hsinstall's mission.

All of this code is isoloted in a new module as well and could serve as
an illustration of how to do this in the future.