Removed code to force cabal file creation
I ran into some strict IO problems with the Monoid instance for IO
(Maybe FilePath) and decided to back off on this feature altogether. The
code issues an error message again that's clearer now.

See this discussion on ways to solve this using Alternative or a custom
newtype wrapper around MaybeT.

https://stackoverflow.com/questions/47120384/keeping-io-lazy-under-append