Changed how resources and AppImage source files are handled
This is major change to the project

Instead of resources coming from ./resources/ and AppImage files coming
from ./util/resources/appimage/, they are now combined into a ./pack/
directory that's copied onto the PREFIX during execution. This is much
more flexible and allows anything to be deployed like additional
binaries, scripts or documentation.

Part of this change includes the <PREFIX>/share/<PROJECT-NAME>/
directory no longer including the version number. This makes it simpler
to "template" target files for this directory down in
pack/share/<PROJECT-NAME>

Other changes here include better error messages in the Except module
when an IOException occurs.

Closes gh-17