Support project that build with either cabal or stack
We try to detect based on the existence of a `stack.yaml` file in the
current directory.

Changes were also made to how the cabal file is searched for and how
hsinstall attempts to force creation of a cabal file if it's not found
at first. Also the error message for cabal file missing failure has been
updated to reflect the new behavior.

The usage was also changed to reflect the new cabal capabilities.

Closes gh-41.