Using Haskell core type Version for version
Switched from String to Distribution.Types.Version which is the type we
get from the package data anyway. No need to convert this into a
human-readable String until it's used at runtime. And no need to wrap
this field of DeploymentInfo into a newtype when it's already a
distinct, useful type.